Trying a secondary path. Using all object base down to RowItems objects.
Thought this might be a good path to try. As I have mentioned much before, this is for theoretical development. If it works how I picture it, then will be used publicly. That being said, this concept is using ROwItem objects. This may make it easier and more manageable to do queries such as SELECT statements.