Added minor functionality change.
Finding best solution to kep elements stored. I added a dynamic attribute [reach-key] that acts like a key value pair, and keeps track of a global counter for applying it. No two elements will have the same key as it falls in DOM order. This may seem problematic in future, so will be looking into this change for unknown/unexpected bugs.